Revert "rename to aiphysim + refactor losses"
