--- aiphysim/models/spacetime.py
@@ -1,6 +1,0 @@
-import torch.nn as nn
-
-
-class SpaceTime(nn.Module):
-    def __init__(self):
-        pass
