+++ config/opts.yaml
@@ -16,25 +16,12 @@ data_folder:
   pendulum: ./datasets/Pendulum
   fluidbox: ./datasets/FluidFlowBox
   attractor: ./datasets/FluidFlowOnAttractor
-  density: ./datasets/Density
-
-dataset_type:
-  discrete: koopman
-  pendulum: koopman
-  fluidbox: koopman
-  attractor: koopman
-  density: density
 
 sequence_length:
   discrete: 51
   pendulum: 51
   fluidbox: 101
   attractor: 121
-  density: 2499
-
-num_shifts:
-  discrete: 30
-  pendulum: 30
 
 limit: # limit the number of samples loaded ; <=0 means no limit
   train: -1
@@ -45,98 +32,72 @@ limit: # limit the number of samples loaded ; <=0 means no limit
 # -----  MODEL  -----
 # -------------------
 
-model_type: koopman
-
 hidden_layers:
   discrete: 2
   pendulum: 2
   attractor: 2
   fluidbox: 2
-  density: 2
 
 hidden_dim:
   discrete: 128
   pendulum: 128
   fluidbox: 128
   attractor: 128
-  density: 256
 
 hidden_layers_aux:
   discrete: 1
   pendulum: 1
   fluidbox: 1
   attractor: 1
-  density: 2
 
 hidden_dim_aux:
   discrete: 64
   pendulum: 64
   fluidbox: 64
   attractor: 64
-  density: 64
 
 latent_dim:
   discrete: 3
   pendulum: 3
   fluidbox: 3
   attractor: 3
-  density: 3
 
-loss_type: koopman # loss processing scheme
-
-losses: # individual loss functions to use
+losses:
   discrete:
-    - reconstruction
-    - latent_evolution
-    - prediction
-    - inf_norm
-    - l2
+    reconstruction: true
+    koopman: true
+    prediction: true
   pendulum:
-    - reconstruction
-    - latent_evolution
-    - prediction
-    - inf_norm
-    - l2
+    reconstruction: true
+    koopman: true
+    prediction: true
   fluidbox:
-    - reconstruction
-    - latent_evolution
-    - prediction
+    reconstruction: true
+    koopman: true
+    prediction: true
   attractor:
-    - reconstruction
-    - latent_evolution
-    - prediction
-  density:
-    - reconstruction
-    - latent_evolution
-    - prediction
-    - inf_norm
-    - l2
+    reconstruction: true
+    koopman: true
+    prediction: true
 
 weights:
   discrete:
-    reconstruction: 0.1
-    latent_evolution: 1
-    prediction: 0.1
-    inf_norm: 1e-7
-    l2: 1e-15
+    reconstruction: 1
+    koopman: 1
+    prediction: 1
   pendulum:
     reconstruction: 1
-    latent_evolution: 1
+    koopman: 1
     prediction: 1
   fluidbox:
     reconstruction: 1
-    latent_evolution: 1
+    koopman: 1
     prediction: 1
   attractor:
     reconstruction: 1
-    latent_evolution: 1
+    koopman: 1
     prediction: 1
-  density:
-    reconstruction: 0.1
-    latent_evolution: 1
-    prediction: 0.1
-    inf_norm: 1e-7
-    l2: 1e-15
+
 # ----------------------
 # -----  TRAINING  -----
 # ----------------------
