+++ scripts/subsample_density_dataset.py
@@ -1,14 +1,13 @@
-import ast
 import sys
+import ast
 from pathlib import Path
-
 import h5py
 import numpy as np
 
 sys.path.append(str(Path(__file__).resolve().parent.parent))
-import minydra  # noqa: E402
+import minydra
 
-from aiphysim.utils import new_unique_path, resolve  # noqa: E402
+from koop.utils import resolve, new_unique_path
 
 
 def dat_to_array(fname, shape=3):
